+++ tests/resources/stfu.php
@@ -1,0 +1,1 @@
+te<?= @$undefined ?>st